requirements.txt
django
requests
djangorestframework
python-dotenv
djangorestframework-simplejwt
factory_boy
# create a virtual environment
#python -m venv genv

# command to install required packages
# pip install -r requirements.txt